+++ app/page.tsx
@@ -1,11 +1,11 @@
 import NavBar from "@/components/Navbar";
 import { TableList } from "@/components/TableList";
-
+import { Suspense } from "react";
 export default function Home() {
   return (
-    <div>
+    <Suspense fallback={<div>Loading...</div>}>
       <NavBar />
       <TableList />
-    </div>
+    </Suspense>
   );
 }
